backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
redis==5.0.1
minio==7.2.0
requests==2.31.0
httpx==0.25.2
beautifulsoup4==4.12.2
dnspython==2.4.2
python-whois==0.8.0
tldextract==5.1.1
idna==3.6
confusable_homoglyphs==3.2.0
sslyze==6.0.0
scikit-learn==1.3.2
lightgbm==4.1.0
pandas==2.1.4
numpy==1.26.2
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
python-socketio==5.10.0
openai==1.3.8
celery==5.3.4
Pillow==10.1.0
playwright==1.40.0